fixed and improved blocky clouds; modified fog, vl and tonemap

+++ shaders/program/final.glsl
@@ -11,7 +11,7 @@ const float wetnessHalflife = 128.0;
 //Optifine Constants//
 /*
 const int colortex0Format = R11F_G11F_B10F; //scene
-const int colortex1Format = RGB16; //translucent
+const int colortex1Format = RGB8; //translucent
 const int colortex2Format = RGBA16; //normals, emissives, specular
 const int colortex4Format = RGBA8; //ao
 const int colortex5Format = RGBA16; //temporal data
